Add initial support for (at least for *NIX) bind() before doing a connect(), allowing the local address to be specified before connecting out.
It's a bit hairy because evutil_socket_connect_() supports
the notion of allocating the FD itself before doing the bind,
so although evutil_socket_bind_() could be called by
consumers before evutil_socket_connect_(), it wouldn't
be 100% correct.  So teach it about also taking the local
socket details and provide the legacy call which doesn't.