Add a call to the bufferevent_sock API to set the local address of a connect() socket.
It's supposed to be called before bufferevent_sock_connect()
so the local address can be set.  It doesn't at all try to
ensure the local address family matches up with the remote
destination address family, so that's up to the programmer
to get correct.  Notably, if you rely on DNS resolution
and don't know the address family up front, there's no way
at the moment to do this correctly.

It also doesn't yet tie into the win32 connect() path.
That should be reasonably trivial to do.